refactor(#142/*): wish success code global/response 로 이동
