+++ src/main/java/org/sopt/snappinserver/global/response/code/wish/WishSuccessCode.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.api.v1.wish.code;
+package org.sopt.snappinserver.global.response.code.wish;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
